nsc-typechecker: move Typechecker monad to its own module
